Create test file per route
